requirements.txt
streamlit
torch
torchvision
opencv-python-headless
pillow
numpy